moved views to views folder
