+++ src/types/index.ts
@@ -1,0 +1,5 @@
+import * as React from "react";
+
+export interface HasToggleFavorites {
+  
+}
