+++ src/views/FavoritesView/FavoritesView.css
