+++ src/views/SelectionView/SelectionView.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import { Button } from "../Button/Button";
+import { Button } from "../../components/Button/Button";
 import "./SelectionView.css";
 
 export interface SelectionViewProps {
